Move the AWS utility packages out of the 'internal' folder
