+++ delta_test.go
@@ -33,13 +33,13 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
 	"github.com/chelseajonesr/rfarrow"
 	"github.com/google/uuid"
-	"github.com/rivian/delta-go/internal/dynamodbutils"
-	"github.com/rivian/delta-go/internal/s3utils"
+	"github.com/rivian/delta-go/dynamodbutils"
 	"github.com/rivian/delta-go/lock"
 	"github.com/rivian/delta-go/lock/dynamolock"
 	"github.com/rivian/delta-go/lock/filelock"
 	"github.com/rivian/delta-go/lock/nillock"
 	"github.com/rivian/delta-go/logstore/dynamodblogstore"
+	"github.com/rivian/delta-go/s3utils"
 	"github.com/rivian/delta-go/state/filestate"
 	"github.com/rivian/delta-go/state/localstate"
