+++ state/dynamostate/dynamostate_test.go
@@ -16,7 +16,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/rivian/delta-go/internal/dynamodbutils"
+	"github.com/rivian/delta-go/dynamodbutils"
 	"github.com/rivian/delta-go/state"
 )
